yida_project/euclid/requirement.txt
django >= 2.1
djangorestframework >= 3.5
djangorestframework-jwt
gunicorn
fcm-django==0.2.19
celery==4.2.0
googlemaps